backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
django-cleanup==8.1.0
django_filter==23.5
gunicorn==21.2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2